backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic-settings==2.6.1
authlib==1.3.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
google-generativeai==0.8.3
fsrs>=6.0.0
pgvector==0.3.6
psycopg2-binary==2.9.10
